requirements.txt
flask
flask-cors
snowflake-connector-python
gunicorn
